fix: side bar border line display issue
The border line was carried by el-scrollbar--view, it either left lots
of blank, or the border line scrolls up when window height is smaller
than menu height. It seems that this can't be fixed by changing
el-scrollbar itself; the line was added by el-menu. We have to move the
border line from el-menu to el-scrollbar manually.